Support only two resolvers resolver, email and without catalog user
Previously Sign-In required a pre-existing catalog User entity.  This
patch adds a resolver to allow users to sign in without an existing User
entity in the catalog.

When users exist in the catalog, they're matched based on their email
from the spec.profile.email field of the entity.

With this patch there is the intentional decision to until support these
two sign in resolvers for the time being to minimize configuration and
support burden.

The name and profile picture are extracted from the id token if the
`name` and `picture` claims are present.