Verify emailMatchingUserEntityProfileEmail works
Previously no User or Group entities are in the catalog because the
production configuration cleared the lost of locations to import data
from.

This patch imports initial User and Group data from the iam.yaml file,
which is present both in the container image and overridden by the
Deployment volumeMount.

For a holos portal, the Secret is expected to be populated from CUE
templates and personalized.  The iam.yaml built into the image will need
to have example.com replace openinfrastructure.co.