Add GitHub Provider
The GitHub integration has a discovery provider for discovering catalog entities within a GitHub organization. The provider will crawl the GitHub organization and register entities matching the configured path. This can be useful as an alternative to static locations or manually adding things to the catalog. This is the preferred method for ingesting entities into the catalog.

You will have to add the GitHub Entity provider to your backend as it is not installed by default, therefore you have to add a dependency on @backstage/plugin-catalog-backend-module-github to your backend package.

```sh
yarn --cwd packages/backend add @backstage/plugin-catalog-backend-module-github
```